LayoutTests/imported/w3c/web-platform-tests/css/css-box/margin-trim/invalidation/block-container-block-start-self-collapsing-nested-no-trim-to-block.html

<!DOCTYPE html>
<html>
<head>
<link rel="author" title="Sammy Gill" href="mailto:sammy.gill@apple.com">
<link rel="help" href="https://drafts.csswg.org/css-box-4/#margin-trim-block">
<meta name="assert" content="Block start margin trimming is performed when property margin-trim goes from none to block">
<style>
container {
    display: block;
    border: 1px solid black;
    margin-block-start: 10px;
}
item {
    display: block;
    inline-size: 50px;
    background-color: green;
}
.collapsed {
    margin-block-start: 50px;
    block-size: 0px
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
</head>
<body onload=" checkLayout('container, item');">
    <container data-expected-margin-top="10">
        <item data-expected-margin-top="0" class="collapsed">
            <item data-expected-margin-top="0" class="collapsed"></item>
        </item>
        <item data-expected-margin-top="0" style="margin-block: 40px">
            <item data-expected-margin-top="0" data-expected-margin-bottom="0" class="collapsed"></item>
            <item data-expected-margin-top="0" style="margin-top: 30px;">
                <item data-expected-margin-top="0" style="margin-block-start: 100px; height: 50px;"></item>
            </item>
        </item>
    </container>
</body>
<script>
document.body.offsetHeight;
document.querySelector("container").style["margin-trim"] = "block";
document.body.offsetHeight;
</script>
</html>
